Add files via upload
Contracted preprojective algebras are studied here: https://arxiv.org/abs/2409.05603. Let A=KQ/I be a preprojective algebra of Dynkin type and e an idempotent given as the sum of primitive idempotents corresponding to the vertices of A. A contracted preprojective algebra of Dynkin type is an algebra Morita equivalent to eAe. For the proof of  Theorem 4.6(4) of the article https://arxiv.org/abs/2409.05603 the homological dimension idim B, gldim B and domdim B can be calculated using QPA when B is a contracted preprojective algebra of Dynkin type E_6. The following text-file shows how to use QPA to obtain explicit quiver and relations of contracted preprojective algebras.